requirements.txt
PyQt5==5.15.10
pyqt5-tools
opencv-python
python-xlib==0.33